requirements.txt
pandas~=1.5.3
scikit-learn~=1.2.1
numpy~=1.24.2
torch~=1.13.1
matplotlib~=3.7.0
torchvision~=0.14.1
tqdm~=4.65.0
Pillow~=9.4.0